requirements.txt
crewai
crewai-tools
langchain-community
langchain-google-genai
langchain-openai
langchain-groq